+++ src/main/java/net/mvndicraft/townyroads/settings/TownyRoadsSettings.java
@@ -124,4 +124,8 @@ public class TownyRoadsSettings {
     public static boolean getRoadsPermissionNationLeadersCanAccept() {
         return Settings.getBoolean(ConfigNodes.ROADS_PERMISSION_NATION_LEADERS_CAN_ACCEPT);
     }
+
+    public static int getMinY() {
+        return Settings.getInt(ConfigNodes.ROADS_OTHER_MIN_Y);
+    }
 }