+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Add different color for valid, blocked & invalid roads #7
 - improve can't build in road message
 - fix circle roads not beeing display as hollow #3
+- Allow players to build/break in roads below y=50 #8
 
 # 1.0.6
 - Make sure road name is aways right.